units/requirements.txt
pytest
pytest-asyncio
psycopg2-binary
python-telegram-bot
sqlalchemy
asyncio
requests
dotenv
